Replace uses of std::string with std::filesystem::path.
In C++, we will ALWAYS (with the exception of passing to/from miniz,
which will correctly handle UTF-8 encoded paths on windows) use
std::filesystem::path to represent local filesystem paths. When we pass
into Lua land, these paths will be converted to UTF-8: both Windows and
POSIX will exhibit the same behavior in Lua, which means mod writers do
not have to care what the underlying platform is.

This will also log UTF-8 encoded strings to the console, which will look
super wrong on Windows, but this just needs a change in the Logger to
perform UTF-8 to UTF-16 conversion.