Make frame recording thread safe, async, and write to disk immediately.
Recording frames asynchronously allows us to not block the render thread (up until the buffer is full, at least). We also cache a bunch of textures that we can reuse to avoid allocating a new one every time, and also write immediately to disk at the first chance we get. We also write BMPs for now instead of PNGs, because PNG encoding time takes too long and wtill stall recording.